ManausNotebookLM-baseline/requirements.txt
Flask
Flask-Cors
Flask-SQLAlchemy
selenium
requests